ipl_chatbot/requirements.txt
flask
langchain
faiss-cpu
pypdf
google-generativeai
python-dotenv
langchain_community
sentence-transformers
langchain_huggingface